requirements.txt
PyYAML
redis
TwitterAPI
